Added Page Buttons to Bottom of Search Page + Lazy Load Images
Added buttons to the bottom of the search page. They function exactly like the ones on the top of the search page. Also updated the a tags for the buttons into actual buttons for accessibility reasons

Added a loading lazy flag to the images on the search page and user list pages as it will allow for some of the images to be loaded later if the user scrolls near them.