Fixed List Entry Bug + Update Styling for Mobile on Book Page
Fixed an issue where if the modal was cancelled, the button to add the entry to the list would update to 'Update List' eventhough no update was sent.

Updated styling on the book page to handle mobile phone & tablet widths better. Most styling changes were applied to the aside div.