Unified Search Endpoints + Fixed Duplicate Issue on Imprint Page
Unified all of the search endpoints to utilize the same, generic one that was built to handle filtering. The generic search endpoint also contains the new search results response with the nextPage boolean flag. Components have been modified to deal with the new object structure.

With the unification, the duplication issue where going back one month and returning cause new requests and events has been fixed. It was likely an issue with the date output but since it's directly built from the date object, the complexity was removed and the bug with it.